requirements.txt
colorama==0.4.6
Jinja2==3.1.6
matplotlib==3.10.3
numpy==2.3.1
pandas==2.3.1
praw==7.8.1
python-dotenv==1.1.1
scikit_learn==1.7.1
seaborn==0.13.2
sentence_transformers==5.0.0
torch==2.7.1+cu128
tqdm==4.67.1
